Bugfix in mesh index buffer

+++ src/mesh.cpp
@@ -36,7 +36,7 @@ Mesh::Mesh(const vector<GLfloat>& vertices, const vector<GLuint> indices) : Mesh
     // IB
     glGenBuffers(1, &indexBuffer);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, indexBuffer);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, indices.size() * sizeof(GLfloat), indices.data(), GL_STATIC_DRAW);
+    glBufferData(GL_ELEMENT_ARRAY_BUFFER, indices.size() * sizeof(GLuint), indices.data(), GL_STATIC_DRAW);
 
     glBindVertexArray(0);
 }
